[CLN] calendar: clean up code and split test files
In this commit, I replaced the old service definition convention with the
updated one. Additionally, I split the test files, moving the
tour-related tests into a separate file for better organization.
