+++ addons/calendar/static/src/views/attendee_calendar/attendee_calendar_model.js
@@ -6,9 +6,11 @@ import { askRecurrenceUpdatePolicy } from "@calendar/views/ask_recurrence_update
 import { deleteConfirmationMessage, ConfirmationDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 
 export class AttendeeCalendarModel extends CalendarModel {
-    setup(params, { dialog }) {
+    static services = [...CalendarModel.services, "dialog", "orm"];
+
+    setup(params, services) {
         super.setup(...arguments);
-        this.dialog = dialog;
+        this.dialog = services.dialog;
         this.rpc = rpc;
     }
 
@@ -206,4 +208,3 @@ export class AttendeeCalendarModel extends CalendarModel {
         await this.load();
     }
 }
-AttendeeCalendarModel.services = [...CalendarModel.services, "dialog", "orm"];
