+++ addons/calendar/tests/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import test_access_rights
@@ -6,6 +5,7 @@ from . import test_attendees
 from . import test_calendar
 from . import test_calendar_controller
 from . import test_calendar_recurrent_event_case2
+from . import test_calendar_tour
 from . import test_event_recurrence
 from . import test_event_notifications
 from . import test_mail_activity_mixin
