+++ addons/calendar/tests/test_recurrence_rule.py
@@ -1,3 +1,5 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from datetime import datetime
 
 from odoo.tests.common import TransactionCase
