+++ addons/calendar/tests/test_res_users.py
@@ -1,3 +1,5 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from odoo.tests.common import TransactionCase
